PracticaClaseAS1 - Test Unitarios/src/Persona.java
public class Persona {
    private String nombre;
    private String apellido;
    private String email;
    private int edad;

    public Persona(String nombre, String apellido, String email, int edad) {
        this.nombre = nombre;
        this.apellido = apellido;
        this.email = email;
        this.edad = edad;
    }

    public String getNombre() {
        return nombre;
    }

    public String getApellido() {
        return apellido;
    }

    public int getEdad() {
        return edad;
    }

    public void mostrarNombre(){
        System.out.println(apellido + ", " + nombre);
    }

    public boolean mayor18anios(){
        return edad >= 18;
    }
}
